refactored with ListItemFragment
