+++ app/src/main/java/com/freshdigitable/udonroad2/MainActivityNavigation.kt
@@ -9,6 +9,7 @@ import com.freshdigitable.udonroad2.navigation.FragmentContainerState
 import com.freshdigitable.udonroad2.navigation.Navigation
 import com.freshdigitable.udonroad2.navigation.NavigationDispatcher
 import com.freshdigitable.udonroad2.navigation.NavigationEvent
+import com.freshdigitable.udonroad2.timeline.ListItemFragment
 import com.freshdigitable.udonroad2.timeline.TimelineEvent
 import com.freshdigitable.udonroad2.timeline.TimelineFragment
 import com.freshdigitable.udonroad2.timeline.TweetDetailFragment
@@ -61,7 +62,7 @@ class MainActivityNavigation(
                 if (isStackedOnTop(BACK_STACK_TWEET_DETAIL)) {
                     activity.supportFragmentManager.popBackStack()
                 } else {
-                    replace(TimelineFragment.newInstance(ListQuery.Timeline()))
+                    replace(ListItemFragment.newInstance<TimelineFragment>(ListQuery.Timeline()))
                 }
                 viewModel.setFabVisible(true)
             }
