+++ feature/timeline/src/main/java/com/freshdigitable/udonroad2/timeline/ListLoadable.kt
@@ -1,0 +1,13 @@
+package com.freshdigitable.udonroad2.timeline
+
+import androidx.lifecycle.LiveData
+import androidx.paging.PagedList
+
+interface ListLoadable {
+    val loading: LiveData<Boolean>
+    fun onRefresh()
+}
+
+interface ListItemLoadable<T> : ListLoadable {
+    fun getList(listOwner: ListOwner): LiveData<PagedList<T>>
+}
