+++ feature/timeline/src/main/res/layout/fragment_timeline.xml
@@ -40,7 +40,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:onRefreshListener="@{viewModel::onRefresh}"
-            app:refreshing="@{safeUnbox(viewModel.loading)}"
+            app:refreshing="@{safeUnbox(viewModel.getLoading())}"
             >
             <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/main_list"
